refractored to add background gradient into helper method functionality
