+++ labs-ios-starter/Dashboard/DashboardViewController.swift
@@ -1,0 +1,32 @@
+//
+//  DashboardViewController.swift
+//  labs-ios-starter
+//
+//  Created by Stephanie Ballard on 12/8/20.
+//  Copyright © 2020 Spencer Curtis. All rights reserved.
+//
+
+import UIKit
+
+class DashboardViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        configureUI()
+    }
+    
+    private func configureUI() {
+        Helper.configureGradientLayer(view: view)
+    }
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
